Widget generator refactoring: we cannot store the widgets anymore
  - Remove widgets endpoints from the SDK
  - New route for widget: /widget-generator
  - Update widget page: no more side panel
