+++ apps/dashboard/src/app/components/navbar/navbar.component.html
@@ -118,7 +118,7 @@
           *ngIf="isWidgetsEnabled | async"
           class="group-items">
           <a
-            [routerLink]="kbUrl + '/widgets'"
+            [routerLink]="kbUrl + '/widget-generator'"
             routerLinkActive="active"
             class="nav-link"
             [class.disabled]="!(isAdminOrContrib | async)">
