--- apps/dashboard/src/app/widgets/add/add-widget.component.html
@@ -1,36 +1,0 @@
-<div class="widget-dialog">
-  <h1>{{ (data?.rename ? 'stash.widgets.rename-title' : 'stash.widgets.add') | translate }}</h1>
-  <pa-button
-    class="close-button"
-    icon="cross"
-    aspect="basic"
-    (click)="close()">
-    {{ 'generic.close' | translate }}
-  </pa-button>
-
-  <form
-    [formGroup]="mainForm"
-    (ngSubmit)="save()">
-    <pa-input
-      formControlName="id"
-      [help]="'stash.widgets.add-help' | translate"
-      [showAllErrors]="false"
-      [errorMessages]="validationMessages">
-      Id
-    </pa-input>
-
-    <div class="buttons">
-      <pa-button
-        aspect="basic"
-        (click)="close()">
-        {{ 'generic.cancel' | translate }}
-      </pa-button>
-      <pa-button
-        type="submit"
-        kind="primary"
-        [disabled]="!mainForm.valid || mainForm.pristine">
-        {{ 'generic.save' | translate }}
-      </pa-button>
-    </div>
-  </form>
-</div>
