--- apps/dashboard/src/app/widgets/add/add-widget.component.scss
@@ -1,17 +1,0 @@
-@import 'variables';
-
-.widget-dialog {
-  position: relative;
-  padding: rhythm(1.5) rhythm(4);
-
-  h1 {
-    margin: 0 0 rhythm(2) 0;
-  }
-
-  .buttons {
-    display: flex;
-    justify-content: center;
-    margin-top: rhythm(2);
-    gap: rhythm(2);
-  }
-}
