--- apps/dashboard/src/app/widgets/edit/edit-widget.component.html
@@ -1,115 +1,0 @@
-<div class="edit-widget">
-  <section>
-    <form
-      [formGroup]="mainForm"
-      (change)="generateSnippet()">
-      <h2>
-        {{ 'stash.widgets.title' | translate }}
-        <span>{{ isDefaultWidget ? 'Dashboard widget' : widget?.id }}</span>
-      </h2>
-      <div class="one-col">
-        <div>
-          <span class="label">{{ 'stash.widgets.preview' | translate }}</span>
-        </div>
-        <div>
-          <div
-            id="preview"
-            [innerHtml]="snippetPreview"></div>
-        </div>
-      </div>
-      <div class="two-cols">
-        <div>
-          <div class="container options">
-            <div class="label">
-              {{ 'stash.widgets.preferences' | translate }}
-            </div>
-
-            <div class="options-content">
-              <pa-select
-                [label]="'stash.widgets.mode' | translate"
-                formControlName="mode">
-                <pa-option value="input">{{ 'stash.widgets.mode.input' | translate }}</pa-option>
-                <pa-option value="form">{{ 'stash.widgets.mode.form' | translate }}</pa-option>
-                <pa-option value="video">{{ 'stash.widgets.mode.video' | translate }}</pa-option>
-              </pa-select>
-              <div
-                class="features"
-                formGroupName="features"
-                *ngIf="widgetMode !== 'video'">
-                <pa-toggle
-                  *ngIf="isTrainingEnabled | async"
-                  formControlName="suggestLabels">
-                  {{ 'stash.widgets.suggestLabels' | translate }}
-                </pa-toggle>
-                <pa-toggle formControlName="editLabels">{{ 'stash.widgets.editLabels' | translate }}</pa-toggle>
-                <pa-toggle formControlName="entityAnnotation">
-                  {{ 'stash.widgets.entityAnnotation' | translate }}
-                </pa-toggle>
-                <pa-toggle formControlName="filter">{{ 'stash.widgets.filter' | translate }}</pa-toggle>
-                <pa-toggle formControlName="permalink">{{ 'stash.widgets.permalink' | translate }}</pa-toggle>
-                <pa-toggle formControlName="navigateToLink">{{ 'stash.widgets.navigateToLink' | translate }}</pa-toggle>
-              </div>
-              <div *ngIf="mainForm.controls.mode.value === 'form'">
-                <pa-input
-                  [(value)]="placeholder"
-                  (valueChange)="onPlaceholderChange($event)">
-                  {{ 'stash.widgets.placeholder' | translate }}
-                </pa-input>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="container snippet">
-          <div class="label">
-            <span>
-              {{ 'stash.widgets.snippet' | translate }}
-            </span>
-            <div class="snippet-buttons">
-              <pa-button
-                *ngIf="widgetMode !== 'input'"
-                aspect="basic"
-                icon="info"
-                iconAndText
-                (click)="showHints()">
-                {{ 'stash.widgets.recommendations' | translate }}
-              </pa-button>
-              <pa-button
-                *ngIf="clipboardSupported"
-                aspect="basic"
-                size="small"
-                [icon]="copyIcon"
-                (click)="copy()">
-                Copy snippet
-              </pa-button>
-            </div>
-          </div>
-          <pre><code class="highlight">{{ snippet }}</code></pre>
-        </div>
-      </div>
-      <div class="buttons">
-        <pa-button
-          kind="primary"
-          [disabled]="(!mainForm.valid || mainForm.pristine) && placeholder === widget?.placeholder"
-          (click)="save()">
-          {{ 'generic.save' | translate }}
-        </pa-button>
-        <pa-button
-          [disabled]="mainForm.pristine"
-          (click)="cancel()">
-          {{ 'generic.cancel' | translate }}
-        </pa-button>
-        <pa-button
-          [disabled]="isDefaultWidget"
-          (click)="rename()">
-          {{ 'stash.widgets.rename' | translate }}
-        </pa-button>
-        <pa-button
-          kind="destructive"
-          [disabled]="isDefaultWidget"
-          (click)="deleteWidget()">
-          {{ 'generic.delete' | translate }}
-        </pa-button>
-      </div>
-    </form>
-  </section>
-</div>
