+++ apps/dashboard/src/app/widgets/widget-features.ts
@@ -1,0 +1,15 @@
+import { WidgetFeatures } from '@nuclia/core';
+
+export const DEFAULT_FEATURES: WidgetFeatures = {
+  editLabels: true,
+  entityAnnotation: true,
+  filter: true,
+  navigateToLink: true,
+  permalink: true,
+  notPublic: true,
+};
+
+export const DEFAULT_FEATURES_LIST = Object.entries(DEFAULT_FEATURES)
+  .filter(([, value]) => value)
+  .map(([key]) => key)
+  .join(',');
