--- apps/dashboard/src/app/widgets/widgets.component.html
@@ -1,25 +1,0 @@
-<stf-section-navbar
-  [title]="'stash.widgets'"
-  [routes]="(routes | async) || []"
-  [hasInnerRoutes]="true">
-  <stf-section-navbar-body class="widgets-sections">
-    <router-outlet></router-outlet>
-  </stf-section-navbar-body>
-  <stf-section-navbar-footer>
-    <pa-button (click)="addWidget()">
-      {{ 'stash.widgets.add' | translate }}
-    </pa-button>
-    <div class="warning">
-      <pa-icon name="warning"></pa-icon>
-      <div>
-        <span [innerHTML]="'stash.widgets.warning' | translate"></span>
-        <ng-container *ngIf="showLink | async">
-          <span>&nbsp;</span>
-          <span [innerHTML]="'stash.widgets.publish_kb' | translate"></span>
-          <span>&nbsp;</span>
-          <a [routerLink]="homeUrl | async">{{ 'stash.widgets.publish_kb_link' | translate }}</a>
-        </ng-container>
-      </div>
-    </div>
-  </stf-section-navbar-footer>
-</stf-section-navbar>
