--- apps/dashboard/src/app/widgets/widgets.component.scss
@@ -1,19 +1,0 @@
-@import 'variables';
-
-.widgets-sections {
-  display: block;
-  margin: 30px 0 80px 0;
-}
-
-.warning {
-  background: $color-secondary-regular;
-  display: flex;
-  gap: rhythm(1);
-  margin-top: rhythm(6);
-  padding: rhythm(1);
-
-  a {
-    color: inherit;
-    text-decoration: underline;
-  }
-}
