+++ apps/dashboard/src/app/widgets/widgets.module.ts
@@ -11,23 +11,15 @@ import {
 } from '@guillotinaweb/pastanaga-angular';
 import { TranslateModule } from '@ngx-translate/core';
 import { AngularSvgIconModule } from 'angular-svg-icon';
-import { AddWidgetDialogModule } from './add/add-widget.module';
-import { EditWidgetComponent } from './edit/edit-widget.component';
+import { WidgetGeneratorComponent } from './widget-generator.component';
 
-import { WidgetsComponent } from './widgets.component';
 import { WidgetHintDialogComponent } from './hint/widget-hint.component';
 import { STFSectionNavbarModule } from '../components/section-navbar';
 
 const routes = [
   {
     path: '',
-    component: WidgetsComponent,
-    children: [
-      {
-        path: ':id',
-        component: EditWidgetComponent,
-      },
-    ],
+    component: WidgetGeneratorComponent,
   },
 ];
 
@@ -41,13 +33,12 @@ const routes = [
     PaTextFieldModule,
     PaButtonModule,
     PaTogglesModule,
-    AddWidgetDialogModule,
     AngularSvgIconModule,
     PaIconModule,
     PaModalModule,
   ],
   exports: [],
-  declarations: [EditWidgetComponent, WidgetsComponent, WidgetHintDialogComponent],
+  declarations: [WidgetGeneratorComponent, WidgetHintDialogComponent],
   providers: [],
 })
 export class WidgetsModule {}
