+++ libs/common/src/assets/i18n/ca.json
@@ -553,14 +553,15 @@
 	"stash.valid_emails_label": "Correus electrònics",
 	"stash.view_edit_service_access": "Visualitza / Edita els Service Access existents",
 	"stash.widgets": "Widgets",
+	"stash.widgets.configuration": "Configuració del widget",
 	"stash.widgets.editLabels": "Edita les etiquetes",
 	"stash.widgets.entityAnnotation": "Anotació de l'entitat",
 	"stash.widgets.filter": "Filtre",
 	"stash.widgets.hint_1": "Per mostrar correctament els resultats, l'amplada de la barra de cerca no hauria de ser inferior a 348px.",
 	"stash.widgets.hint_2": "Si el color de fons no és neutral, recorda d'afegir un color a la caixa de cerca per fer-la més llegible.",
 	"stash.widgets.mode": "Mode",
-	"stash.widgets.mode.form": "Giny de cerca incrustat",
-	"stash.widgets.mode.input": "Giny de cerca emergent",
+	"stash.widgets.mode.embedded": "Giny de cerca incrustat",
+	"stash.widgets.mode.popup": "Giny de cerca emergent",
 	"stash.widgets.mode.video": "Barra de cerca i widget de vídeo",
 	"stash.widgets.navigateToLink": "Navega cap als enllaços",
 	"stash.widgets.permalink": "Enllaços permenents",
