+++ libs/common/src/assets/i18n/en.json
@@ -645,6 +645,7 @@
 	"stash.widgets": "Widgets",
 	"stash.widgets.add": "Add search widget",
 	"stash.widgets.add-help": "Name your widget to save it",
+	"stash.widgets.configuration": "Widget configuration",
 	"stash.widgets.editLabels": "Edit labels",
 	"stash.widgets.entityAnnotation": "Entity annotation",
 	"stash.widgets.filter": "Filter",
@@ -652,13 +653,12 @@
 	"stash.widgets.hint_2": "If the background you'll be adding your search isn't neutral, remember to add a background for your search box to make it easier to read.",
 	"stash.widgets.invalid-id": "Invalid widget id",
 	"stash.widgets.mode": "Mode",
-	"stash.widgets.mode.form": "Embedded search widget",
-	"stash.widgets.mode.input": "Popup search widget",
+	"stash.widgets.mode.embedded": "Embedded search widget",
+	"stash.widgets.mode.popup": "Popup search widget",
 	"stash.widgets.mode.video": "Search bar and video widget",
 	"stash.widgets.navigateToLink": "Navigate to links",
 	"stash.widgets.permalink": "Permalinks",
 	"stash.widgets.placeholder": "Placeholder",
-	"stash.widgets.preferences": "Preferences",
 	"stash.widgets.preview": "Preview",
 	"stash.widgets.publish_kb": "You can publish it from the",
 	"stash.widgets.publish_kb_link": "Dashboard",
