+++ libs/common/src/assets/i18n/es.json
@@ -553,14 +553,15 @@
 	"stash.valid_emails_label": "Correos electrónicos",
 	"stash.view_edit_service_access": "Visualiza / Edita los Service Access existentes",
 	"stash.widgets": "Widgets",
+	"stash.widgets.configuration": "Configuracion de Widget",
 	"stash.widgets.editLabels": "Editar etiquetas",
 	"stash.widgets.entityAnnotation": "Anotación de entidad",
 	"stash.widgets.filter": "Filtrar",
 	"stash.widgets.hint_1": "Para mostrar correctamente los resultados, el ancho de la barra de búsqueda no debería ser inferior a 384px.",
 	"stash.widgets.hint_2": "Si el color de fondo no es neutral, recuerda añadir un color a la caja de búsqueda para que sea más legible.",
 	"stash.widgets.mode": "Modo",
-	"stash.widgets.mode.form": "Widget de búsqueda integrado",
-	"stash.widgets.mode.input": "Widget de búsqueda emergente",
+	"stash.widgets.mode.embedded": "Widget de búsqueda integrado",
+	"stash.widgets.mode.popup": "Widget de búsqueda emergente",
 	"stash.widgets.mode.video": "Barra de búsqueda y widget de video",
 	"stash.widgets.navigateToLink": "Navega hacia los enlaces",
 	"stash.widgets.permalink": "Enlaces permanentes",
