+++ libs/sdk-core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nuclia/core",
-  "version": "1.0.6",
+  "version": "1.1.0",
   "description": "SDK allowing to integrate Nuclia services in your frontend application",
   "license": "MIT",
   "keywords": [
