+++ libs/search-widget/src/widgets/search-widget/Widget.svelte
@@ -26,7 +26,6 @@
   import { initViewerEffects, unsubscribeViewerEffects } from '../../core/old-stores/viewer-effects';
 
   export let backend = 'https://nuclia.cloud/api';
-  export let widgetid = '';
   export let zone = '';
   export let knowledgebox = '';
   export let type = 'embedded'; // 'popup' | 'embedded'
@@ -77,7 +76,6 @@
 
   onMount(() => {
     initNuclia(
-      widgetid,
       {
         backend,
         zone,
